Jormungandr: small refacto for the default scenario
Externalization of helpers function like the pagination
